# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_http-server_handlers_auth.Credentials:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  backend_internal_http-server_handlers_auth.singUpResponse:
    properties:
      message:
        type: string
    type: object
  backend_internal_http-server_handlers_auth.tokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  backend_internal_lib_response.errorResponse:
    description: This structure is used to send error information in JSON format.
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse:
    description: This structure is used to send error information in JSON format.
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    name: GitHub
    url: https://github.com/AmadoMuerte
  description: API для sysStats, платформа для сбора предпочтений пользователей
    при выборе фильма для совместного просмотра.
  title: sysStats API
  version: "1.0"
paths:
  /login/refresh:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to refresh their access token using
        a valid refresh token.
      parameters:
      - description: Bearer {refresh_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_http-server_handlers_auth.tokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
      summary: Refresh Access token
      tags:
      - Authentication
  /login/sign-in:
    post:
      consumes:
      - application/json
      description: This endpoint allows user to sign in using their email and passwd.
      parameters:
      - description: Credentials for signing in
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/backend_internal_http-server_handlers_auth.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_http-server_handlers_auth.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
      summary: Sign In
      tags:
      - Authentication
  /login/sign-up:
    post:
      consumes:
      - application/json
      description: This endpoint allows user to sign up using their email and passwd.
      parameters:
      - description: Credentials for signing up
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/backend_internal_http-server_handlers_auth.Credentials'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_http-server_handlers_auth.singUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AmadoMuerte_sysStats_internal_lib_response.errorResponse'
      summary: Sign Up
      tags:
      - Authentication
schemes:
- http
- https
swagger: "2.0"